requirements.txt
biopython
numpy
pandas
matplotlib
seaborn
scikit-learn
jupyter
